updated appconfig, moved client code
