+++ lib/client/supabase-client.ts
@@ -1,5 +1,5 @@
 import { createClient } from "@supabase/supabase-js";
-import { storageAdapter } from "./storage";
+import { storageAdapter } from "../storage";
 
 const supabaseUrl = process.env.EXPO_PUBLIC_SUPABASE_URL;
 const supabaseAnonKey = process.env.EXPO_PUBLIC_SUPABASE_ANON_KEY;
